+++ audiorep/domain/__init__.py
@@ -6,6 +6,7 @@ from audiorep.domain.label import Label
 from audiorep.domain.playlist import Playlist, PlaylistEntry
 from audiorep.domain.cd_disc import CDDisc, CDTrack, RipStatus
 from audiorep.domain.radio_station import RadioStation
+from audiorep.domain.eq_preset import EqPreset
 
 __all__ = [
     "Track", "AudioFormat", "TrackSource",
@@ -15,4 +16,5 @@ __all__ = [
     "Playlist", "PlaylistEntry",
     "CDDisc", "CDTrack", "RipStatus",
     "RadioStation",
+    "EqPreset",
 ]